+++ home/tests/test_api.py
@@ -491,6 +491,7 @@ class TestContentPageAPI:
         assert meta["slug"] == page.slug
         assert meta["parent"]["id"] == page.get_parent().id
         assert meta["locale"] == "en"
+        assert meta["detail_url"] == f"http://localhost/api/v2/pages/{page.id}/"
 
         assert content == {
             "id": page.id,
@@ -502,6 +503,42 @@ class TestContentPageAPI:
             "quick_replies": [],
             "related_pages": [],
             "has_children": False,
+        }
+
+    @pytest.mark.parametrize("platform", ALL_PLATFORMS)
+    def test_detail_view_meta(self, uclient, platform):
+        """
+        Fetching the detail view of a page returns the page metadata.
+        """
+        page = self.create_content_page(tags=["self_help"], body_type=platform)
+        response = uclient.get(f"/api/v2/pages/{page.id}/")
+        content = response.json()
+
+        # There's a lot of metadata, so only check selected fields.
+        meta = content.pop("meta")
+        parent = page.get_parent()
+
+        assert meta == {
+            "type": "home.ContentPage",
+            "detail_url": f"http://localhost/api/v2/pages/{page.id}/",
+            "html_url": page.get_full_url(),
+            "slug": page.slug,
+            "show_in_menus": "false",
+            "seo_title": page.seo_title,
+            "search_description": page.search_description,
+            "first_published_at": page.first_published_at.strftime(
+                "%Y-%m-%dT%H:%M:%S.%fZ"
+            ),
+            "alias_of": page.alias_of,
+            "parent": {
+                "id": parent.id,
+                "meta": {
+                    "type": "home.ContentPageIndex",
+                    "html_url": parent.get_full_url(),
+                },
+                "title": parent.title,
+            },
+            "locale": "en",
         }
 
     def test_detail_view_increments_count(self, uclient):
@@ -540,6 +577,7 @@ class TestContentPageAPI:
         assert meta["slug"] == page.slug
         assert meta["parent"]["id"] == page.get_parent().id
         assert meta["locale"] == "en"
+        assert meta["detail_url"] == f"http://localhost/api/v2/pages/{page.id}/"
 
         assert content["has_children"] is True
 
@@ -559,6 +597,7 @@ class TestContentPageAPI:
         assert meta["slug"] == page.slug
         assert meta["parent"]["id"] == page.get_parent().id
         assert meta["locale"] == "en"
+        assert meta["detail_url"] == f"http://localhost/api/v2/pages/{page.id}/"
 
         assert content["id"] == page.id
         assert content["title"] == "default page"
@@ -605,6 +644,7 @@ class TestContentPageAPI:
         assert meta["slug"] == page.slug
         assert meta["parent"]["id"] == page.get_parent().id
         assert meta["locale"] == "en"
+        assert meta["detail_url"] == f"http://localhost/api/v2/pages/{page.id}/"
 
         assert content["id"] == page.id
         assert content["title"] == "default page"
@@ -1499,6 +1539,10 @@ class TestAssessmentAPI:
         assert meta["slug"] == self.high_result_page.slug
         assert meta["parent"]["id"] == self.high_result_page.get_parent().id
         assert meta["locale"] == "en"
+        assert (
+            meta["detail_url"]
+            == f"http://localhost/api/v2/pages/{self.high_result_page.id}/"
+        )
         assert content["results"][0]["high_result_page"] == {
             "id": self.high_result_page.id,
             "title": self.high_result_page.title,
@@ -1517,6 +1561,10 @@ class TestAssessmentAPI:
         assert meta["slug"] == self.medium_result_page.slug
         assert meta["parent"]["id"] == self.medium_result_page.get_parent().id
         assert meta["locale"] == "en"
+        assert (
+            meta["detail_url"]
+            == f"http://localhost/api/v2/pages/{self.medium_result_page.id}/"
+        )
         assert content["results"][0]["medium_result_page"] == {
             "id": self.medium_result_page.id,
             "title": self.medium_result_page.title,
@@ -1535,6 +1583,10 @@ class TestAssessmentAPI:
         assert meta["slug"] == self.low_result_page.slug
         assert meta["parent"]["id"] == self.low_result_page.get_parent().id
         assert meta["locale"] == "en"
+        assert (
+            meta["detail_url"]
+            == f"http://localhost/api/v2/pages/{self.low_result_page.id}/"
+        )
         assert content["results"][0]["low_result_page"] == {
             "id": self.low_result_page.id,
             "title": self.low_result_page.title,
@@ -1552,6 +1604,10 @@ class TestAssessmentAPI:
         assert meta["slug"] == self.skip_high_result_page.slug
         assert meta["parent"]["id"] == self.skip_high_result_page.get_parent().id
         assert meta["locale"] == "en"
+        assert (
+            meta["detail_url"]
+            == f"http://localhost/api/v2/pages/{self.skip_high_result_page.id}/"
+        )
         assert content["results"][0]["skip_high_result_page"] == {
             "id": self.skip_high_result_page.id,
             "title": self.skip_high_result_page.title,
@@ -1655,6 +1711,10 @@ class TestAssessmentAPI:
         assert meta["slug"] == self.high_result_page.slug
         assert meta["parent"]["id"] == self.high_result_page.get_parent().id
         assert meta["locale"] == "en"
+        assert (
+            meta["detail_url"]
+            == f"http://localhost/api/v2/pages/{self.high_result_page.id}/"
+        )
         assert content["high_result_page"] == {
             "id": self.high_result_page.id,
             "title": self.high_result_page.title,
@@ -1673,6 +1733,10 @@ class TestAssessmentAPI:
         assert meta["slug"] == self.medium_result_page.slug
         assert meta["parent"]["id"] == self.medium_result_page.get_parent().id
         assert meta["locale"] == "en"
+        assert (
+            meta["detail_url"]
+            == f"http://localhost/api/v2/pages/{self.medium_result_page.id}/"
+        )
         assert content["medium_result_page"] == {
             "id": self.medium_result_page.id,
             "title": self.medium_result_page.title,
@@ -1691,6 +1755,10 @@ class TestAssessmentAPI:
         assert meta["slug"] == self.low_result_page.slug
         assert meta["parent"]["id"] == self.low_result_page.get_parent().id
         assert meta["locale"] == "en"
+        assert (
+            meta["detail_url"]
+            == f"http://localhost/api/v2/pages/{self.low_result_page.id}/"
+        )
         assert content["low_result_page"] == {
             "id": self.low_result_page.id,
             "title": self.low_result_page.title,
@@ -1723,6 +1791,10 @@ class TestAssessmentAPI:
         assert meta["slug"] == self.high_result_page.slug
         assert meta["parent"]["id"] == self.high_result_page.get_parent().id
         assert meta["locale"] == "en"
+        assert (
+            meta["detail_url"]
+            == f"http://localhost/api/v2/pages/{self.high_result_page.id}/"
+        )
         assert content["results"][0]["high_result_page"] == {
             "id": self.high_result_page.id,
             "title": self.high_result_page.title,
@@ -1741,6 +1813,10 @@ class TestAssessmentAPI:
         assert meta["slug"] == self.medium_result_page.slug
         assert meta["parent"]["id"] == self.medium_result_page.get_parent().id
         assert meta["locale"] == "en"
+        assert (
+            meta["detail_url"]
+            == f"http://localhost/api/v2/pages/{self.medium_result_page.id}/"
+        )
         assert content["results"][0]["medium_result_page"] == {
             "id": self.medium_result_page.id,
             "title": self.medium_result_page.title,
@@ -1759,6 +1835,10 @@ class TestAssessmentAPI:
         assert meta["slug"] == self.low_result_page.slug
         assert meta["parent"]["id"] == self.low_result_page.get_parent().id
         assert meta["locale"] == "en"
+        assert (
+            meta["detail_url"]
+            == f"http://localhost/api/v2/pages/{self.low_result_page.id}/"
+        )
         assert content["results"][0]["low_result_page"] == {
             "id": self.low_result_page.id,
             "title": self.low_result_page.title,
@@ -1804,6 +1884,10 @@ class TestAssessmentAPI:
         assert meta["slug"] == self.high_result_page.slug
         assert meta["parent"]["id"] == self.high_result_page.get_parent().id
         assert meta["locale"] == "en"
+        assert (
+            meta["detail_url"]
+            == f"http://localhost/api/v2/pages/{self.high_result_page.id}/"
+        )
         assert content["high_result_page"] == {
             "id": self.high_result_page.id,
             "title": self.high_result_page.title,
@@ -1822,6 +1906,10 @@ class TestAssessmentAPI:
         assert meta["slug"] == self.medium_result_page.slug
         assert meta["parent"]["id"] == self.medium_result_page.get_parent().id
         assert meta["locale"] == "en"
+        assert (
+            meta["detail_url"]
+            == f"http://localhost/api/v2/pages/{self.medium_result_page.id}/"
+        )
         assert content["medium_result_page"] == {
             "id": self.medium_result_page.id,
             "title": self.medium_result_page.title,
@@ -1840,6 +1928,10 @@ class TestAssessmentAPI:
         assert meta["slug"] == self.low_result_page.slug
         assert meta["parent"]["id"] == self.low_result_page.get_parent().id
         assert meta["locale"] == "en"
+        assert (
+            meta["detail_url"]
+            == f"http://localhost/api/v2/pages/{self.low_result_page.id}/"
+        )
         assert content["low_result_page"] == {
             "id": self.low_result_page.id,
             "title": self.low_result_page.title,
@@ -1852,18 +1944,17 @@ class TestAssessmentAPI:
             "triggers": [],
         }
 
-    # TODO: Add this test in once we've merged with main - there's some code in there that makes following the redirect work
-    # def test_assessment_detail_endpoint_without_drafts(self, uclient, settings):
-    #     """
-    #     Unpublished assessments are not returned if the qa param is not set.
-    #     """
-    #     settings.STATIC_ROOT = Path("home/tests/test_static")
-    #     self.assessment.unpublish()
-    #     url = f"/api/v2/assessment/{self.assessment.id}"
+    def test_assessment_detail_endpoint_without_drafts(self, uclient, settings):
+        """
+        Unpublished assessments are not returned if the qa param is not set.
+        """
+        settings.STATIC_ROOT = Path("home/tests/test_static")
+        self.assessment.unpublish()
+        url = f"/api/v2/assessment/{self.assessment.id}"
 
-    #     response = uclient.get(url, follow=True)
+        response = uclient.get(url, follow=True)
 
-    #     assert response.status_code == 404
+        assert response.status_code == 404
 
     def test_assessment_new_draft(self, uclient):
         """
@@ -1906,6 +1997,10 @@ class TestAssessmentAPI:
         assert meta["slug"] == self.high_result_page.slug
         assert meta["parent"]["id"] == self.high_result_page.get_parent().id
         assert meta["locale"] == "en"
+        assert (
+            meta["detail_url"]
+            == f"http://localhost/api/v2/pages/{self.high_result_page.id}/"
+        )
         assert content["results"][0]["high_result_page"] == {
             "id": self.high_result_page.id,
             "title": self.high_result_page.title,
@@ -1924,6 +2019,10 @@ class TestAssessmentAPI:
         assert meta["slug"] == self.medium_result_page.slug
         assert meta["parent"]["id"] == self.medium_result_page.get_parent().id
         assert meta["locale"] == "en"
+        assert (
+            meta["detail_url"]
+            == f"http://localhost/api/v2/pages/{self.medium_result_page.id}/"
+        )
         assert content["results"][0]["medium_result_page"] == {
             "id": self.medium_result_page.id,
             "title": self.medium_result_page.title,
@@ -1942,6 +2041,10 @@ class TestAssessmentAPI:
         assert meta["slug"] == self.low_result_page.slug
         assert meta["parent"]["id"] == self.low_result_page.get_parent().id
         assert meta["locale"] == "en"
+        assert (
+            meta["detail_url"]
+            == f"http://localhost/api/v2/pages/{self.low_result_page.id}/"
+        )
         assert content["results"][0]["low_result_page"] == {
             "id": self.low_result_page.id,
             "title": self.low_result_page.title,
@@ -1963,6 +2066,10 @@ class TestAssessmentAPI:
         assert meta["slug"] == high_result_page.slug
         assert meta["parent"]["id"] == high_result_page.get_parent().id
         assert meta["locale"] == "en"
+        assert (
+            meta["detail_url"]
+            == f"http://localhost/api/v2/pages/{high_result_page.id}/"
+        )
         assert content["results"][0]["high_result_page"] == {
             "id": high_result_page.id,
             "title": high_result_page.title,
@@ -1981,6 +2088,10 @@ class TestAssessmentAPI:
         assert meta["slug"] == self.medium_result_page.slug
         assert meta["parent"]["id"] == self.medium_result_page.get_parent().id
         assert meta["locale"] == "en"
+        assert (
+            meta["detail_url"]
+            == f"http://localhost/api/v2/pages/{self.medium_result_page.id}/"
+        )
         assert content["results"][0]["medium_result_page"] == {
             "id": self.medium_result_page.id,
             "title": self.medium_result_page.title,
@@ -1999,6 +2110,10 @@ class TestAssessmentAPI:
         assert meta["slug"] == self.low_result_page.slug
         assert meta["parent"]["id"] == self.low_result_page.get_parent().id
         assert meta["locale"] == "en"
+        assert (
+            meta["detail_url"]
+            == f"http://localhost/api/v2/pages/{self.low_result_page.id}/"
+        )
         assert content["results"][0]["low_result_page"] == {
             "id": self.low_result_page.id,
             "title": self.low_result_page.title,
@@ -2022,6 +2137,10 @@ class TestAssessmentAPI:
         assert meta["slug"] == self.high_result_page.slug
         assert meta["parent"]["id"] == self.high_result_page.get_parent().id
         assert meta["locale"] == "en"
+        assert (
+            meta["detail_url"]
+            == f"http://localhost/api/v2/pages/{self.high_result_page.id}/"
+        )
         assert content["results"][0]["high_result_page"] == {
             "id": self.high_result_page.id,
             "title": self.high_result_page.title,
@@ -2040,6 +2159,10 @@ class TestAssessmentAPI:
         assert meta["slug"] == self.medium_result_page.slug
         assert meta["parent"]["id"] == self.medium_result_page.get_parent().id
         assert meta["locale"] == "en"
+        assert (
+            meta["detail_url"]
+            == f"http://localhost/api/v2/pages/{self.medium_result_page.id}/"
+        )
         assert content["results"][0]["medium_result_page"] == {
             "id": self.medium_result_page.id,
             "title": self.medium_result_page.title,
@@ -2058,6 +2181,10 @@ class TestAssessmentAPI:
         assert meta["slug"] == self.low_result_page.slug
         assert meta["parent"]["id"] == self.low_result_page.get_parent().id
         assert meta["locale"] == "en"
+        assert (
+            meta["detail_url"]
+            == f"http://localhost/api/v2/pages/{self.low_result_page.id}/"
+        )
         assert content["results"][0]["low_result_page"] == {
             "id": self.low_result_page.id,
             "title": self.low_result_page.title,
@@ -2084,6 +2211,10 @@ class TestAssessmentAPI:
         assert meta["slug"] == self.high_result_page.slug
         assert meta["parent"]["id"] == self.high_result_page.get_parent().id
         assert meta["locale"] == "en"
+        assert (
+            meta["detail_url"]
+            == f"http://localhost/api/v2/pages/{self.high_result_page.id}/"
+        )
         assert content["high_result_page"] == {
             "id": self.high_result_page.id,
             "title": self.high_result_page.title,
@@ -2102,6 +2233,10 @@ class TestAssessmentAPI:
         assert meta["slug"] == self.medium_result_page.slug
         assert meta["parent"]["id"] == self.medium_result_page.get_parent().id
         assert meta["locale"] == "en"
+        assert (
+            meta["detail_url"]
+            == f"http://localhost/api/v2/pages/{self.medium_result_page.id}/"
+        )
         assert content["medium_result_page"] == {
             "id": self.medium_result_page.id,
             "title": self.medium_result_page.title,
@@ -2120,6 +2255,10 @@ class TestAssessmentAPI:
         assert meta["slug"] == self.low_result_page.slug
         assert meta["parent"]["id"] == self.low_result_page.get_parent().id
         assert meta["locale"] == "en"
+        assert (
+            meta["detail_url"]
+            == f"http://localhost/api/v2/pages/{self.low_result_page.id}/"
+        )
         assert content["low_result_page"] == {
             "id": self.low_result_page.id,
             "title": self.low_result_page.title,